ImagePlusCreator: combine Color- and GrayImagePlusCreator
ColorImagePlusCreator and GrayImagePlusCreator have only little
differences. The code duplication is removed by merging them together.
